requirements.txt
config==0.3.9
dj-database-url==0.5.0

django-bootstrap3==10.0.1
django-bootstrap4==0.0.8
django-heroku==0.3.1
django-tinymce==2.8.0
gunicorn==19.9.0
Pillow==5.2.0
psycopg2==2.7.5
python-decouple==3.1
pytz==2018.5
whitenoise==4.0
